增加了部分牌的立绘显示效果。 修复了多人游戏时遇见特定怪物时游戏闪退的异常。 修复了多人游戏时遭遇战事件无法正常开始的异常。 修改了【|账号已封禁|】，【《四号屠场 · 众生安眠》】，【冠军晚餐·猫的摇篮】等牌的效果。 修复了一些遗物的异常。移除了陨石球。 修复了背景音乐重叠的异常。 修复了扑满在重新加载时卡牌改变的异常。

+++ src/main/java/hsrmod/monsters/Exordium/SilvermaneLieutenant.java
@@ -20,7 +20,7 @@ public class SilvermaneLieutenant extends BaseMonster implements PreBreakSubscri
     int block = 10;
     int thornCount = 3;
     
-    public SilvermaneLieutenant(int x, int y) {
+    public SilvermaneLieutenant(float x, float y) {
         super(ID, 0F, -15.0F, 300, 384, x, y);
         setDamagesWithAscension(7);
         block = ModHelper.moreHPAscension(type) ? 10 : 8;
